backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend FastAPI + Uvicorn"
authors = [
    {name = "Giuseppe GRASSO",email = "giugrasso@outlook.be"},
    {name = "Julien CABERGS",email = "cabergsj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "ollama (>=0.5.4,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
